+++ NEWS.md
@@ -1,9 +1,14 @@
+# QTLseqr 0.5.4
+
+## Updates
+* `plotGprimeDist` now plots histograms of filtered and raw data. overlayed with the null dist. Is easier to interpret. 
+
 # QTLseqr 0.5.3
 
 ## New features
 
 * Added a `NEWS.md` file to track changes to the package.
-* in `getGprimeDist` plotting now includes density plots of all data, data after QTL filtering and the null-distribution assuming mean and variance of the filtered set.
+* in `plotGprimeDist` plotting now includes density plots of all data, data after QTL filtering and the null-distribution assuming mean and variance of the filtered set.
 
 ## Bug fixes
 * Corrected a bug in `getFDRThreshold` function that was using regular p-values and not adjusted pvalues to define threshold